Extract out ModelStub, StorageStub
Since it'll be reused in AddRoomCommandTest
