+++ src/test/java/seedu/resireg/logic/commands/CommandTestUtil.java
@@ -1,7 +1,6 @@
 package seedu.resireg.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.resireg.logic.parser.CliSyntax.PREFIX_ALIAS;
 import static seedu.resireg.logic.parser.CliSyntax.PREFIX_COMMAND;
@@ -18,31 +17,24 @@ import static seedu.resireg.testutil.Assert.assertThrows;
 import static seedu.resireg.testutil.TypicalStudents.AMY;
 import static seedu.resireg.testutil.TypicalStudents.BOB;
 
-import java.io.IOException;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 import seedu.resireg.commons.core.index.Index;
-import seedu.resireg.commons.exceptions.DataConversionException;
 import seedu.resireg.logic.CommandHistory;
 import seedu.resireg.logic.commands.exceptions.CommandException;
 import seedu.resireg.model.Model;
-import seedu.resireg.model.ReadOnlyResiReg;
-import seedu.resireg.model.ReadOnlyUserPrefs;
 import seedu.resireg.model.ResiReg;
-import seedu.resireg.model.UserPrefs;
 import seedu.resireg.model.bin.Binnable;
 import seedu.resireg.model.room.Room;
 import seedu.resireg.model.room.RoomNameContainsKeywordPairsPredicate;
 import seedu.resireg.model.student.NameContainsKeywordsPredicate;
 import seedu.resireg.model.student.Student;
-import seedu.resireg.storage.Storage;
 import seedu.resireg.testutil.EditRoomDescriptorBuilder;
 import seedu.resireg.testutil.EditStudentDescriptorBuilder;
+import seedu.resireg.testutil.StorageStub;
 
 /**
  * Contains helper methods for testing commands.
@@ -265,55 +257,5 @@ public class CommandTestUtil {
         Student firstStudent = model.getFilteredStudentList().get(0);
         model.deleteStudent(firstStudent);
         model.saveStateResiReg();
-    }
-
-    /**
-     * A stub class for Storage.
-     */
-    private static class StorageStub implements Storage {
-        @Override
-        public Path getUserPrefsFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Optional<UserPrefs> readUserPrefs() throws DataConversionException, IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void saveUserPrefs(ReadOnlyUserPrefs userPrefs) throws IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getResiRegFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Optional<ReadOnlyResiReg> readResiReg() throws DataConversionException, IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Optional<ReadOnlyResiReg> readResiReg(Path filePath) throws DataConversionException, IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void saveResiReg(ReadOnlyResiReg resiReg) throws IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void saveResiReg(ReadOnlyResiReg resiReg, Path filePath) throws IOException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void archiveResiReg(ReadOnlyResiReg resiReg) throws IOException {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 }
